integrate telegram highjack into story
